refactor: 랜덤유틸 클래스 분리
